Start refactoring quotas domain

+++ fin-life-api/database/migrations/1681066339531-create_buys_sells_table.ts
@@ -28,7 +28,7 @@ export class createBuysSellsTable1681066339531 implements MigrationInterface {
           },
           {
             name: 'price',
-            type: 'float'
+            type: 'int'
           },
           {
             name: 'type',
